Fix issues in calculation and update test expectations
This commit fixes some issues found in the calculations applied to various physics models. Also, in response to these updates, the corresponding unit tests have been updated to reflect the correct values that are expected to be returned. This includes updating the test values in the coordinates and precession tests, as well as adjustments to improve the precision of calculations in several other areas of the codebase.