Remove earth model calculations, introduce OpenCL support and kernel functions
This commit removes a set of calculations related to the earth model and adds support for OpenCL (Open Computing Language). It includes an OpenCL context along with associated methods for detecting an OpenCL installation on the user's system. Additionally, it incorporates functions for handling OpenCL kernels, which will aid in executing code more efficiently on OpenCL-compatible devices.