Add sun position calculation and corresponding constants
This commit adds a new file 'position.c' in the sun model directory that includes the necessary functions to calculate the sun's position. Also, it includes the required constants in 'constants.h' and 'constants.c' for the calculations. Changes have also been made in 'setup.py' to include this new extension module for package distribution.

Slightly off from Vallado algorithm 29 however I used more precise up to date values for anomaly and solar mean longitude so I'm not pressed for it.