Remove debug output and refine magnitude calculation in sun_position
This commit eliminates a debug printf statement that was present in the sun_position function. Additionally, it improves precision in the calculation of the sun's position by using more accurate constants in the calculation of the magnitude. These changes bring our implementation closer to the Vallado algorithm, even though it's slightly varying due to updated values.