Sun and moon positions (#22)
* Add sun position calculation and corresponding constants

This commit adds a new file 'position.c' in the sun model directory that includes the necessary functions to calculate the sun's position. Also, it includes the required constants in 'constants.h' and 'constants.c' for the calculations. Changes have also been made in 'setup.py' to include this new extension module for package distribution.

Slightly off from Vallado algorithm 29 however I used more precise up to date values for anomaly and solar mean longitude so I'm not pressed for it.

* Remove debug output and refine magnitude calculation in sun_position

This commit eliminates a debug printf statement that was present in the sun_position function. Additionally, it improves precision in the calculation of the sun's position by using more accurate constants in the calculation of the magnitude. These changes bring our implementation closer to the Vallado algorithm, even though it's slightly varying due to updated values.

* Add moon position calculation and related constants

This commit introduces a new feature to compute the position of the moon. It adds two new constants for lunar longitude and parallax, and creates a new model for the moon's position. New files position.h and position.c are added for the moon position calculations. Further, the setup.py is updated to include these new files in the 'toluene_extensions.models.moon.position' extension.

* Making Ernest W Brown proud by ignoring most of his fin precision calculations and just approximating his work.